% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{buildkey}
\alias{buildkey}
\title{build key for key data frame}
\usage{
buildkey(kkey, decode, isSE)
}
\arguments{
\item{decode}{The decoding number if the values have to be interpreted by the CT_SystemConst sheet.}

\item{isSE}{NA if the variable is not in the sub-sheet. Otherwise, look up the SE sub-sheet and fill in the TypeCode of the variable.}

\item{key}{The column name.}
}
\value{
A single row in the data frame format. Rbind the keys to one single data frame and apply it in the the EE_getETSfromJY function.
}
\description{
build key for key data frame
}

