% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{daily2monthly}
\alias{daily2monthly}
\title{Transform daily data to montly data.}
\usage{
daily2monthly(df, date, value, std = TRUE, tradingday = TRUE)
}
\arguments{
\item{df}{The data frame containing the daily data.}

\item{date}{The column of date.}

\item{value}{The column of factorscore.}

\item{std}{Logical values, whether normalize the factorscore.}

\item{tradingday}{Logical values, whether apply trday.nearest function.}
}
\value{
A TSF object.
}
\description{
Transform daily data to montly data.
}

