% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_day0}
\alias{EE_day0}
\title{conduct day 0 statistics analysis}
\usage{
EE_day0(ets)
}
\arguments{
\item{ets}{ets.}
}
\value{
A list containing suspend_ratio, winning_raio and stat_table.
}
\description{
conduct day 0 statistics analysis
}
\examples{
ets <- ets.employeeplan()
EE_day0(ets)
}

