% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_distribute}
\alias{EE_distribute}
\title{plot event histogram}
\usage{
EE_distribute(ETS, TSErr, bin = c("year", "yearmonth", "month", "yearquarter",
  "quarter"))
}
\description{
plot event histogram
}

