% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_ExpandETS_1row}
\alias{EE_ExpandETS_1row}
\title{Expand ETS object into TS object with index}
\usage{
EE_ExpandETS_1row(ETS, win1 = 20, win2 = 60)
}
\arguments{
\item{ETS}{A event TS object which includes the event date and the corresponding stock.}

\item{win1}{Integer. The time window of days before the event date.}

\item{win2}{Integer. The time window of days after the event date.}

\item{adjtrday}{Logical values. Whether to adjust the event date to the next nearest trading day.}
}
\value{
A TS object with index.
}
\description{
Expand ETS object into TS object with index
}

