% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_getETSfromJY}
\alias{EE_getETSfromJY}
\title{Get ETS object from JY database}
\usage{
EE_getETSfromJY(SheetName, key.df, extra.condition = NULL,
  stock.column = "InnerCode", stock.decode = "InnerCode",
  date.column = "InfoPublDate")
}
\arguments{
\item{SheetName}{The sheet name string of the data.}

\item{key.df}{The key data frame.}

\item{extra.condition}{The extra qr string could be add to the SQL.}

\item{stock.column}{The column name which indicating the stockID in the sheet}

\item{stock.decode}{The column name in decode sheet which is used to decode the stockID.}

\item{date.column}{The column name string of the date in the data.}
}
\value{
A dataframe with date, stockID and vars.
}
\description{
Get ETS object from JY database
}
\examples{
# Fetching the ETS of investors' activities.
key.df <- rbind(buildkey("SerialNb"))
tmpdat <- EE_getETSfromJY(SheetName = "LC_InvestorRa", key.df)
# Fetching the ETS of shares transfering.
key.df <- rbind(buildkey("TranShareType",1040),
                buildkey("TranMode",1202),
                buildkey("IfSuspended"),
                buildkey("TransfererEcoNature",1096,1),
                buildkey("ReceiverEcoNature",1096,2))
tmpdat <- EE_getETSfromJY(SheetName = "LC_ShareTransfer", key.df)
}

