% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_getTSErr}
\alias{EE_getTSErr}
\title{Plug in ETS and return a TS object with Err and index.}
\usage{
EE_getTSErr(ETS, err_database = c("EE_CroxSecReg", "EE_CroxSecGroup",
  "pct_chg", "pct_chg_bmk"), win1 = 10, win2 = 20)
}
\arguments{
\item{ETS}{A event TS object which includes the event date and the corresponding stock.}

\item{win1}{Integer. The time window of days before the event date.}

\item{win2}{Integer. The time window of days after the event date.}

\item{db}{The name string of local database err sheet which containing the columns of "date", "stockID", "err".}

\item{cleansing}{Logical values. If true, the result will be fill NA and winsorized before out put.}
}
\value{
A TS object with Err and index.
}
\description{
Plug in ETS and return a TS object with Err and index.
}
\examples{
date <- as.Date(c("2014-07-01","2014-07-08"))
stockID <- c("EQ000001","EQ000002")
ETS <- data.frame(date, stockID)
TSErr <- EE_getTSErr(ETS)
EE_plot(TSErr)
}

