% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_SplitAnalyzer}
\alias{EE_SplitAnalyzer}
\title{split ets into group and conduct analysis}
\usage{
EE_SplitAnalyzer(ets, group_mode, customized_level, db = c("EE_CroxSecReg",
  "pct_chg"), win1 = 20, win2 = 30, withplot = TRUE,
  minimum_group_num = 0)
}
\arguments{
\item{ets}{ETS.}

\item{group_mode}{One single character string. Could be "year", "sector", "simple_sector", "EQ"(002,300,etc.), "size".}

\item{customized_level}{A vector indicating how to group ets. If this argument is supplied, the group_mode will be neglect.}

\item{db}{The database of studying objects. Whether to study residuals or daily_pct_chg.}

\item{win1}{The time window(days) before the events happened.}

\item{win2}{The time window(days) after the events happened.}

\item{withplot}{Logical value.}

\item{minimum_group_num}{The minimum number for each sub group.}
}
\value{
list.
}
\description{
split ets into group and conduct analysis
}
\examples{
ets <- ets.employeeplan()
EE_SplitAnalyzer(ets,"year")
EE_SplitAnalyzer(ets,"sector", minimum_group_num = 10)
EE_SplitAnalyzer(ets,"simple_sector")
EE_SplitAnalyzer(ets,"EQ")
EE_SplitAnalyzer(ets,"size")
}

