% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_splitTSErr}
\alias{EE_splitTSErr}
\title{split TSerr}
\usage{
EE_splitTSErr(TSErr, group_mode = c("year", "sector", "EQ"),
  sectorAttr = NULL)
}
\value{
TSErrs, TSErr with sector
}
\description{
split TSerr
}

