% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_splityear}
\alias{EE_splityear}
\title{Split and return the plots of each year}
\usage{
EE_splityear(TSErr, everyyear = FALSE, bmk = NULL)
}
\arguments{
\item{TSErr}{The TSErr object.}

\item{everyyear}{Logical value. Whether to return the plot of details in each year.}

\item{breakwindow}{Logical value. Whether to keep the event window complete or forcely cut the object by year. Default is false.}
}
\value{
A list containing plots and data.
}
\description{
Split and return the plots of each year
}

