% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_table}
\alias{EE_table}
\title{Plug in TSErr object and return the mean summary table}
\usage{
EE_table(TSErr, withSD = TRUE, withACC = FALSE)
}
\arguments{
\item{TSErr}{The TSErr object}

\item{withSD}{Logical. Whether to compute sd.}

\item{withACC}{Logical. Whether to compute accumulated err.}
}
\value{
Dataframe.
}
\description{
Plug in TSErr object and return the mean summary table
}

