% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_wrap_analyzer}
\alias{EE_wrap_analyzer}
\title{Wrap up ETS analyzing functions}
\usage{
EE_wrap_analyzer(ets, err_database = c("EE_CroxSecReg", "EE_CroxSecGroup",
  "pct_chg", "pct_chg_bmk"), win1 = 10, win2 = 20)
}
\arguments{
\item{win1}{The time window(days) before the events happened.}

\item{win2}{The time window(days) after the events happened.}

\item{db}{The database of studying objects. Whether to study residuals or daily_pct_chg.}
}
\value{
list. Containing stat table, day0 analyzing, daily perfomance and plots.
}
\description{
Wrap up ETS analyzing functions
}
\examples{
ets <- ets.employeeplan()
EE_wrap_analyzer(ets)
}

