% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{ets.EQ002_forecast}
\alias{ets.EQ002_forecast}
\title{get ETS of EQ002 forecast strategy.}
\usage{
ets.EQ002_forecast(ob_win = 20, withlatest = FALSE)
}
\value{
ETS object.
}
\description{
get ETS of EQ002 forecast strategy.
}

