% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{ets.leader_trade}
\alias{ets.leader_trade}
\title{get ETS of leader trading stocks.}
\usage{
ets.leader_trade(direction = c("buy", "sell"))
}
\value{
ETS object.
}
\description{
get ETS of leader trading stocks.
}

