% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{ets.leaderbuy_largebuy}
\alias{ets.leaderbuy_largebuy}
\title{get ETS of leader buying stocks a lot within a few times.}
\usage{
ets.leaderbuy_largebuy()
}
\value{
ETS object.
}
\description{
get ETS of leader buying stocks a lot within a few times.
}

