% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{ets.leadersell_largesell}
\alias{ets.leadersell_largesell}
\title{get ETS of leader selling stocks a lot within a few times.}
\usage{
ets.leadersell_largesell()
}
\value{
ETS object.
}
\description{
get ETS of leader selling stocks a lot within a few times.
}

