% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{ets.port_backtest}
\alias{ets.port_backtest}
\title{back test function for ets}
\usage{
ets.port_backtest(ets, wgt_limit = 0.1, fee.buy = 0, fee.sell = 0)
}
\arguments{
\item{ets}{the ets object must contain the columns date, stockID and date_end.}

\item{wgt_limit}{the biggest wgt of each stock.}
}
\value{
Same result type with port.backtest
}
\description{
the function is modified by port.backtest. This revised version could provide different date_end for each stock. The weight is set to eq for each trday. The date in ets is the buying date and the pct_chg in that date is not obtained. The date_end is the selling date and the pct_chg is included.
}

