% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{ets.unfreeze}
\alias{ets.unfreeze}
\title{get ETS of stocks unfreezing}
\usage{
ets.unfreeze(withPercent = FALSE, research_mode = FALSE)
}
\value{
ETS object.
}
\description{
get ETS of stocks unfreezing
}

