% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{etstack}
\alias{etstack}
\title{plug in ets and return accumulated value of certain varialbe.}
\usage{
etstack(etsobj, win1 = 20)
}
\arguments{
\item{etsobj}{An etsobj.}

\item{win1}{The days window before the event to calculate the accumulated value.}
}
\value{
etsobj
}
\description{
plug in ets and return accumulated value of certain varialbe.
}

