% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{etstick}
\alias{etstick}
\title{plug in ets and return frequency of certain period.}
\usage{
etstick(ets, win1 = 20)
}
\arguments{
\item{ets}{A ets object.}

\item{win1}{The days window before the event to calculate the frequency.}
}
\value{
etsobj
}
\description{
plug in ets and return frequency of certain period.
}

