% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{fillfacNA}
\alias{fillfacNA}
\title{Fill in the NA in TSF}
\usage{
fillfacNA(tsf, trim = NA)
}
\arguments{
\item{tsf}{A tsf object.}
}
\value{
tsf without NA in factor.
}
\description{
Fill in the NA in TSF
}

