% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladder_position_strategy.R
\name{getBondrtn}
\alias{getBondrtn}
\title{Get bond daily return from WindR}
\usage{
getBondrtn(startdate = "2009-01-01", enddate = "2016-06-30",
  code = "037.CS")
}
\value{
return bond daily return from WindR
}
\description{
Get bond daily return from WindR
}

