% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{getETS}
\alias{getETS}
\title{plug in ts object and return with ets object.}
\usage{
getETS(tsobj, EventSet = NULL, win = 20)
}
\arguments{
\item{tsobj}{A ts object.}

\item{EventSet}{a vector to specify the events.}
}
\value{
A data frame with T,S,E and diff.
}
\description{
plug in ts object and return with ets object.
}

