% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{getETSscore}
\alias{getETSscore}
\title{plug in ts object and return with ets score.}
\usage{
getETSscore(tsobj)
}
\arguments{
\item{tsobj}{A tsobj.}
}
\value{
ts object with event score.
}
\description{
plug in ts object and return with ets score.
}

