% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{getmonthFac}
\alias{getmonthFac}
\title{Transform daily TSF to montly TSF.}
\usage{
getmonthFac(ts, db, window = NA)
}
\arguments{
\item{ts}{The TS object.}

\item{db}{The data frame containing the daily TSF data.}

\item{window}{An interger indicating the time window to summarize. If NA, the time window will be the time interval between the rebalance dates.}
}
\value{
A TSF object.
}
\description{
Transform daily TSF to montly TSF.
}

