% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{gf.ezsec}
\alias{gf.ezsec}
\title{Get simple sector}
\usage{
gf.ezsec(ts)
}
\arguments{
\item{ts}{The object containing ts.}
}
\value{
Return the object with additional column secID.
}
\description{
Get simple sector
}
\details{
simple sector only contains 6 categories, ES1:Big Cycle, ES2:Finance and Real estate, ES3:TMT, ES4:Consumpstions, ES5:Manufactoring, ES6:Others.
}

