% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{gf.old_NP_YOY}
\alias{gf.old_NP_YOY}
\title{get old NP YOY}
\usage{
gf.old_NP_YOY(TS, is1q = TRUE, filt = 1e+07, rm_neg = FALSE,
  src = c("all", "fin", "for"), clear_result = TRUE)
}
\description{
get old NP YOY
}

