% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladder_position_strategy.R
\name{LadderNAV}
\alias{LadderNAV}
\title{Function for calculate NAV for Ladder Position Strategy}
\usage{
LadderNAV(rtn, pos = c(0.15, 0.25, 0.35, 0.5), cutvalue = c(1.05, 1.1, 1.2))
}
\arguments{
\item{rtn}{A xts of daily return(value not in percentage).}

\item{pos}{A vector indicating the position values for each stage.}

\item{cutvalue}{A vector indicating the maximum NAV for each stage.}
}
\value{
Returning a data frame, containing return, postion, NAV_1, NAV_2 for each trading day.

The NAV_1 represents the net asset value in the 2-years period, which means the NAV_1 is set to 1.00 at the beginning of every 2 years.

The NAV_2 represents the net asset value since the first date.
}
\description{
Function for calculate NAV for Ladder Position Strategy
}

