% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{lcdb.build.EE_employeeplan}
\alias{lcdb.build.EE_employeeplan}
\title{lcdb.build.EE_employeeplan}
\usage{
lcdb.build.EE_employeeplan()
}
\description{
lcdb.build.EE_employeeplan
}
\examples{
library(WindR)
lcdb.build.EE_employeeplan()
}

