% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{merge2tsf}
\alias{merge2tsf}
\title{Easy way to add up two TSF objects.}
\usage{
merge2tsf(tsf1, tsf2, wgt = c(0.5, 0.5))
}
\arguments{
\item{tsf1, tsf2}{The TSF objects.}

\item{wgt}{The vector of wights which must sum up to 1.}
}
\value{
A TSF object.
}
\description{
Easy way to add up two TSF objects.
}

