% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{MonthMA}
\alias{MonthMA}
\title{get month MA for TS}
\usage{
MonthMA(ts, N)
}
\arguments{
\item{ts}{TS.}
}
\value{
dataframe.
}
\description{
get month MA for TS
}
\examples{
ts <- ets.employeeplan()
N = 9
test <- MonthMA(ts,N)
}

