% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{paramsensi}
\alias{paramsensi}
\title{test sensitivity of the parameters}
\usage{
paramsensi(funcqr, param_1, param_2, withPlot = TRUE)
}
\arguments{
\item{funcqr}{The querry as character strings. Replace the param with the name param_1. And plug in the values into later arguments.}

\item{param_1}{The values to test, could be vector or list.}

\item{param_2}{The values to test, could be vector or list.}

\item{withPlot}{Logical value. whether plot the matrix.}
}
\value{
Result matrix.
}
\description{
The function right now could only support 2 params varing. To write the funcqr querry, replace the parameters with the name param_1 and param_2, and plug the vector or list that you would like to test into the latter arguments param_1 and param_2. The result of funcqr could only be one single value.
}
\examples{
paramsensi("sum(1:param_1)", param_1 = 1:20)
paramsensi("param_1*param_2", param_1 = 1:10, param_2 = 1:9)
}

