% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{rm_unbuyable}
\alias{rm_unbuyable}
\title{filter TS that can not be purchased}
\usage{
rm_unbuyable(TS, uplimit = 0.095, BuyDate = c("nextday", "today"),
  datasrc = defaultDataSRC())
}
\arguments{
\item{uplimit}{the cut point}
}
\value{
TS
}
\description{
filter TS that can not be purchased
}

