% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{rm_untradable}
\alias{rm_untradable}
\title{filter stocks that can not be bought or sold}
\usage{
rm_untradable(TS, limit = 0.095, TradeDir = c("buy", "sell"),
  TradeDate = c("nextday", "today"), datasrc = defaultDataSRC())
}
\description{
filter stocks that can not be bought or sold
}

