% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{rms.all}
\alias{rms.all}
\title{filter stocks to remove negative event.}
\usage{
rms.all(tsobj)
}
\arguments{
\item{tsobj}{An object contains ts.}
}
\value{
tsobj.
}
\description{
filter stocks to remove negative event.
}

