% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{rms.low_F_NP}
\alias{rms.low_F_NP}
\title{fitler stocks that have low NP forcast}
\usage{
rms.low_F_NP(ts)
}
\arguments{
\item{ts}{A ts object.}
}
\value{
ts A ts object.
}
\description{
fitler stocks that have low NP forcast
}

