% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{rms.unfroz}
\alias{rms.unfroz}
\title{filter stocks that are going to be unfrozen.}
\usage{
rms.unfroz(tsobj)
}
\arguments{
\item{tsobj}{An object contains ts.}
}
\value{
tsobj.
}
\description{
filter stocks that are going to be unfrozen.
}

