% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{rptDate.forecastInfo}
\alias{rptDate.forecastInfo}
\title{get forecast report info according to rptTS}
\usage{
rptDate.forecastInfo(rptTS)
}
\value{
data frame
}
\description{
get forecast report info according to rptTS
}

