% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{rsrs_timing}
\alias{rsrs_timing}
\title{RSRS timing}
\usage{
rsrs_timing(indexID = "EI000300", begT = as.Date("2002-01-05"),
  endT = Sys.Date() - 1, param_N = 18, param_M = 600, param_S = 0.8)
}
\arguments{
\item{param_N}{The length of time series used to fit beta.}

\item{param_M}{The length of data used to smooth and adjust beta.}

\item{param_S}{The epsilon used to control buy and sell.}
}
\value{
A list, containing rtn and finalre.
}
\description{
RSRS timing
}
