% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{stock2index}
\alias{stock2index}
\title{plug in stockID and return indexID}
\usage{
stock2index(stocklist)
}
\arguments{
\item{stocklist}{vector.}
}
\value{
indexlist, vector.
}
\description{
plug in stockID and return indexID
}

