% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{strategy_rtn_rough}
\alias{strategy_rtn_rough}
\title{turning ets to rtn series}
\usage{
strategy_rtn_rough(etsfunc, begT = as.Date("2013-01-04"), endT = Sys.Date()
  - 1, univ = NULL, ob_win = 20, wgtmax = 0.05)
}
\arguments{
\item{etsfunc}{ETS function strings, supporting multiple functions. The length of the vector must match the length of the argument ob_win.}

\item{begT}{begT}

\item{endT}{endT. If NULL, then using latest trading day.}

\item{univ}{stock universe. If NULL, considering all stocks. Default NULL.}

\item{ob_win}{The holding days window for each ETS strategies.}

\item{wgtmax}{The max weight limit for each stock.}
}
\value{
A list, rtn and port.
}
\description{
A function used to turning ets to strategies and roughly calculating returns.
}
\examples{
list1 <- strategy_rtn_rough(etsfunc = "ets.EQ002_forecast", begT = as.Date("2013-01-04"), ob_win = 20)
ggplot.WealthIndex(list1$rtn)
list2 <- strategy_rtn_rough(etsfunc = c("ets.EQ002_forecast","ets.employeeplan"), begT = as.Date("2013-01-04"), ob_win = c(20,20))
ggplot.WealthIndex(list2$rtn)
}

