% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{strategy_st_init}
\alias{strategy_st_init}
\title{ST strategy initiate}
\usage{
strategy_st_init(begT, endT, wgt_limit = 0.1)
}
\arguments{
\item{wgt_limit}{Default 0.1, the maximum weight for one individual stock.}
}
\value{
A list, containing rtn and port.
}
\description{
ST strategy initiate
}
\examples{
st_strat_relist <- strategy_st_init(begT = as.Date("2010-01-04"), endT = as.Date("2015-04-23"))
}

