% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{subsetCol}
\alias{subsetCol}
\title{adjusted subset function for Chinese string.}
\usage{
subsetCol(tmpdat, colchar, subsetcode)
}
\arguments{
\item{tmpdat}{The data frame.}

\item{colchar}{The column names.}

\item{subsetcode}{The code for the subset that is going to be filtered.}
}
\value{
data frame subset.
}
\description{
adjusted subset function for Chinese string.
}

