% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{tidytsf}
\alias{tidytsf}
\title{Organize and standardize the TSF object.}
\usage{
tidytsf(tsf, reorder = NULL)
}
\arguments{
\item{tsf}{A data frame containg the following columns: date, stocks, factorscore.}

\item{reorder}{The vector indicating the index of date/stock/factorscore.}
}
\value{
A TSF object with standard format.
}
\description{
Organize and standardize the TSF object.
}

