% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{trday.count.vec}
\alias{trday.count.vec}
\title{Count trading days trough vectors.}
\usage{
trday.count.vec(begTvec, endTvec)
}
\arguments{
\item{begTvec}{begT vector.}

\item{endTvec}{endT vector.}
}
\value{
vector.
}
\description{
Count trading days trough vectors.
}

