% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{trday.get.df}
\alias{trday.get.df}
\title{Get tradingday through data frame.}
\usage{
trday.get.df(T.df)
}
\arguments{
\item{T.df}{A data frame with two columns: begT, endT.}
}
\value{
A data frame with only one column: date.
}
\description{
Get tradingday through data frame.
}

