% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{trday.get.vec}
\alias{trday.get.vec}
\title{Get tradingday by passing in vector inputs.}
\usage{
trday.get.vec(begTvec, endTvec)
}
\arguments{
\item{begTvec}{begT vector.}

\item{endTvec}{endT vector.}
}
\value{
A vector that row-bind all the trading days between begT and endT.
}
\description{
Get tradingday by passing in vector inputs.
}

