% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{wraptsfun}
\alias{wraptsfun}
\title{Wrap up tsRemoteCallFunc with additional StockID column}
\usage{
wraptsfun(funchar, funpar = NULL, syspar = NULL)
}
\arguments{
\item{funchar}{A character string indicating which function to use in tsRemoteCallFunc.}

\item{funpar}{A list for function parameters.}

\item{syspar}{A list for system parameters, eg. StockID.}
}
\value{
A complicated list without unlist.
}
\description{
Wrap up tsRemoteCallFunc with additional StockID column
}

