% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_Day0}
\alias{EE_Day0}
\title{conduct day 0 statistics analysis}
\usage{
EE_Day0(ets)
}
\arguments{
\item{ets}{ets.}
}
\value{
A list.
}
\description{
conduct day 0 statistics analysis
}
\examples{
ets <- ets.employeeplan()
EE_Day0(ets)
}

