% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_ExpandETS}
\alias{EE_ExpandETS}
\title{Expand ETS object into TS object with index. Multi-row version.}
\usage{
EE_ExpandETS(ets, win1, win2)
}
\arguments{
\item{ets}{A event TS object which includes the event date and the corresponding stock.}

\item{win1}{Integer. The time window of days before the event date.}

\item{win2}{Integer. The time window of days after the event date.}
}
\value{
A TS object with index.
}
\description{
Expand ETS object into TS object with index. Multi-row version.
}

