% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_getETSfromJY}
\alias{EE_getETSfromJY}
\title{Get ETS object from JY database}
\usage{
EE_getETSfromJY(stock.column = "InnerCode", stock.decode = "InnerCode",
  date.column = "InfoPublDate", SheetName, key.column, key.decode = NULL)
}
\arguments{
\item{stock.column}{The column name string of the stockID in the data.}

\item{stock.decode}{The column name string that is used to decode the stockID.}

\item{date.column}{The column name string of the date in the data.}

\item{SheetName}{The sheet name string of the data.}

\item{key.column}{The column name string of the key variable in the data.}

\item{key.decode}{The decoding number that is used to interpret the key.column.}
}
\value{
A dataframe with date, stockID and var(key column).
}
\description{
Get ETS object from JY database
}
\examples{
# Fetching the ETS of investors' activities.
ETS <- EE_getETSfromJY(stock.column = "InnerCode", stock.decode = "InnerCode",
                       date.column = "InfoPublDate", SheetName = "LC_InvestorRa", key.column = "SerialNb")

}

