% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_GetTSErr}
\alias{EE_GetTSErr}
\title{Plug in ETS and return a TS object with Err and index.}
\usage{
EE_GetTSErr(ETS, db = "EE_CroxSecReg", win1 = 20, win2 = 60)
}
\arguments{
\item{ETS}{A event TS object which includes the event date and the corresponding stock.}

\item{db}{The name string of local database err sheet which containing the columns of "date", "stockID", "err".}

\item{win1}{Integer. The time window of days before the event date.}

\item{win2}{Integer. The time window of days after the event date.}
}
\value{
A TS object with Err and index.
}
\description{
Plug in ETS and return a TS object with Err and index.
}

