% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_Plot}
\alias{EE_Plot}
\title{Plug in TSErr object and return the summary plot}
\usage{
EE_Plot(TSErr)
}
\arguments{
\item{TSErr}{The TSErr object which must containing No and err columns.}
}
\value{
Two plots.
}
\description{
Plug in TSErr object and return the summary plot
}

