% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{EE_table}
\alias{EE_table}
\title{Plug in TSErr object and return the summary table}
\usage{
EE_table(TSErr)
}
\arguments{
\item{TSErr}{The TSErr object}
}
\value{
Table.
}
\description{
Plug in TSErr object and return the summary table
}

