% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{fillna}
\alias{fillna}
\title{Fill in the NA.}
\usage{
fillna(vec, method = "mean", trim = NA)
}
\arguments{
\item{vec}{A vector.}

\item{method}{The method to fill in the NA.}
}
\value{
A vector without NA values.
}
\description{
Fill in the NA.
}

