% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{strategy_st_upd}
\alias{strategy_st_upd}
\title{ST strategy update}
\usage{
strategy_st_upd(st_strat_relist, wgt_limit = 0.1)
}
\arguments{
\item{wgt_limit}{Default 0.1, the maximum weight for one individual stock.}
}
\value{
A list, containing rtn and port.
}
\description{
update the st dataset to the latest date(sys.date - 1).
}
\examples{
st_strat_relist <- strategy_st_init(begT = as.Date("2010-01-04"), endT = as.Date("2015-04-23"), save_result = TRUE)
st_strat_relist <- strategy_st_upd(st_strat_relist)
}

