% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{trday.count.vec}
\alias{trday.count.vec}
\title{Count trading days by passing vector inputs.}
\usage{
trday.count.vec(begTvec, endTvec)
}
\arguments{
\item{begTvec}{begT vector.}

\item{endTvec}{endT vector.}
}
\value{
A vector which each row represents the lag of days between begT and endT.
}
\description{
Count trading days by passing vector inputs.
}

