% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{trday.get.vec}
\alias{trday.get.vec}
\title{Get tradingday through vectors.}
\usage{
trday.get.vec(begTvec, endTvec)
}
\arguments{
\item{begTvec}{begT vector.}

\item{endTvec}{endT vector.}
}
\value{
vector
}
\description{
Get tradingday through vectors.
}

