% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funset.R
\name{wraptsfun}
\alias{wraptsfun}
\title{Wrap up tsRemoteCallFunc with additional StockID column}
\usage{
wraptsfun(funchar, StockID)
}
\arguments{
\item{funchar}{A character string indicating which function to use in tsRemoteCallFunc.}

\item{StockID}{A vector.}
}
\value{
A complicated list without unlist.
}
\description{
Wrap up tsRemoteCallFunc with additional StockID column
}

